requirements.txt
fastapi==0.85.2
flatbuffers==2.0.7
fonttools==4.37.1
Jinja2==3.1.2
matplotlib==3.5.3
numpy==1.23.2
pandas==1.4.4
requests-oauthlib==1.3.1
rsa==4.9
uvicorn==0.19.0
tqdm==4.64.1